Changed files structuration for Sass, moved everything footer related into a Footer component, started styling ListMessages component
