+++ frontend/src/Pages/Home.scss
@@ -1,0 +1,28 @@
+
+.home {
+    display: grid;
+    grid-template-columns: 1fr 2fr 1fr;
+    grid-template-rows: 1fr;
+    grid-column-gap: 0px;
+    grid-row-gap: 0px;
+    height: 60vh;
+    /* background-color: #eca9a0; */
+    border-radius: 10px;
+  }
+  
+  .home-room {
+    grid-area: 1 / 1 / 2 / 2;
+    /* background-color: blue; */
+  }
+    
+  .home-tchat { 
+    grid-area: 1 / 2 / 2 / 3;
+    /* background-color: white; */
+    /* border: red 10px solid; */
+    overflow: auto; 
+  }
+  .home-listplayer { 
+    grid-area: 1 / 3 / 2 / 4;
+    /* background-color: red; */
+    overflow: auto; 
+  }
